+ chg: get dark mode for Statusbar segments only

+++ src/Dialogs.h
@@ -128,8 +128,10 @@ void DeleteBitmapButton(HWND hwnd, int nCtrlId);
 
 
 #define StatusSetSimple(hwnd,b) SendMessage(hwnd,SB_SIMPLE,(WPARAM)b,0)
-void StatusSetText(HWND hwnd, UINT nPart, LPCWSTR lpszText);
-bool StatusSetTextID(HWND hwnd, UINT nPart, UINT uID);
+void StatusSetText(HWND hwnd, BYTE nPart, LPCWSTR lpszText);
+bool StatusSetTextID(HWND hwnd, BYTE nPart, UINT uID);
+//void StatusPartSetText(HWND hwnd, BYTE nPart, LPCWSTR lpszText);
+//bool StatusPartSetTextID(HWND hwnd, BYTE nPart, UINT uID);
 
 int Toolbar_GetButtons(HANDLE hwnd, int cmdBase, LPWSTR lpszButtons, int cchButtons);
 int Toolbar_SetButtons(HANDLE, int, LPCWSTR, void*, int);
